feat: rename SharedComponents folder to Components
