+++ src/Analysis/GWASResults/Views/Home/Home.jsx
@@ -3,7 +3,7 @@ import { Spin } from 'antd';
 import { useQuery } from 'react-query';
 import HomeTable from './HomeTable/HomeTable';
 import { gwasWorkflowPath } from '../../../../localconf';
-import LoadingErrorMessage from '../../SharedComponents/LoadingErrorMessage/LoadingErrorMessage';
+import LoadingErrorMessage from '../../Components/LoadingErrorMessage/LoadingErrorMessage';
 import ManageColumns from './ManageColumns/ManageColumns';
 
 const Home = () => {
